MAINT: add eval_block util to altair.utils
